通Pageコンポーネントを作成
